requirements/app.txt
# THIS FILE CONTAIN
# DEPENDENCIES FOR DEMO ALTROO
# CHAT APPLICATION

Django==3.0.2


# ALTROO CHAT DEPENEDENCIES
channels==2.4.0
channels-redis==2.4.1
Pillow==7.0.0

# DJANGO REST FRAMEWORK
# DEPENDENCIES

djangorestframework==3.11.0
django-filter==2.2.0
Markdown==3.1.1
psycopg2==2.8.4
